+++ views/details.ejs
@@ -1,8 +1,8 @@
 <h2><%= pokemon.name%></h2>
 
-<ul>
-	<li>weight: <%= pokemon.weight%></li>
-	<li>height: <%= pokemon.height%></li>
-	<li>base_experience: <%= pokemon.base_experience%></li>
+<ul class="list-group">
+	<li class="list-group-item">weight: <%= pokemon.weight%></li>
+	<li class="list-group-item">height: <%= pokemon.height%></li>
+	<li class="list-group-item">base_experience: <%= pokemon.base_experience%></li>
 	<img src="<%=pokemon.sprites.back_default%>" alt="picture here">
 </ul>
